Actualizacion de Backend y FrontEnd

+++ backend/src/main/java/com/proof/service/EstudianteService.java
@@ -9,24 +9,53 @@ import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
 
+/**
+ * Servicio de la entidad Estudiante
+ * 
+ * @autor David Orlando Velez Zamora
+ */
 @Service
 public class EstudianteService {
 
     @Autowired
     private EstudianteRepository estudianteRepository;
 
+    /**
+     * Listar todos los estudiantes
+     * 
+     * @return Lista de estudiantes
+     */
     public List<Estudiante> listarEstudiantes() {
         return estudianteRepository.findAll();
     }
 
+    /**
+     * Obtener un estudiante por ID
+     * 
+     * @param id ID del estudiante a obtener
+     * @return Estudiante
+     */
     public Optional<Estudiante> obtenerEstudiantePorId(Long id) {
         return estudianteRepository.findById(id);
     }
 
+    /**
+     * Guardar un estudiante
+     * 
+     * @param estudiante Estudiante a guardar
+     * @return Estudiante
+     */
     public Estudiante guardarEstudiante(Estudiante estudiante) {
         return estudianteRepository.save(estudiante);
     }
 
+    /**
+     * Actualizar un estudiante
+     * 
+     * @param id                    ID del estudiante a actualizar
+     * @param estudianteActualizado Estudiante actualizado
+     * @return Estudiante
+     */
     public Estudiante actualizarEstudiante(Long id, Estudiante estudianteActualizado) {
         return estudianteRepository.findById(id)
                 .map(estudiante -> {
@@ -36,8 +65,19 @@ public class EstudianteService {
                 }).orElseThrow(() -> new RuntimeException("Estudiante no encontrado con ID: " + id));
     }
 
-    public void eliminarEstudiante(Long id) {
-        estudianteRepository.deleteById(id);
+    /**
+     * Eliminar un estudiante
+     * 
+     * @param id ID del estudiante a eliminar
+     * @return boolean
+     */
+    public boolean eliminarEstudiante(Long id) {
+        if (estudianteRepository.existsById(id)) {
+            estudianteRepository.deleteById(id);
+            return true;
+        } else {
+            return false;
+        }
     }
 
 }
